+++ RELEASE_NOTES.md
@@ -9,5 +9,6 @@
 
 - Update the radio button navigation for streamlit in deepnote.
 - Add KWK aliases in the line plot.
+- Update the KPIs to german format.
 
 ## Bug Fixes